+++ packages/application/src/_index.ts
@@ -1,6 +1,6 @@
 export * from './api/_index'
 export * from './dto/_index'
 export * from './actions/_index'
-
+export * from './mockRadix'
 export * from './_types'
 export * from './radix'