+++ packages/application/src/actions/_index.ts
@@ -1,9 +1,4 @@
 export * from './_types'
-
-export * from './otherAction'
-export * from './executedStakeTokensAction'
-export * from './executedStakeTokensAction'
-export * from './executedStakeTokensAction'
 export * from './intendedStakeTokensAction'
 export * from './intendedTransferTokensAction'
 export * from './intendedUnstakeTokensAction'