+++ packages/primitives/src/_index.ts
@@ -1,9 +1,7 @@
 export * from './_types'
-
 export * from './uint256-extensions'
 export * from './resultAsync'
 export * from './magic'
 export * from './amount'
-export * from './nonce'
 export * from './granularity'
 export * from './denomination'