+++ packages/application/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.15](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/application@1.0.14...@radixdlt/application@1.0.15) (2021-04-12)
+
+**Note:** Version bump only for package @radixdlt/application
+
+
+
+
+
+## [1.0.14](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/application@1.0.13...@radixdlt/application@1.0.14) (2021-04-09)
+
+**Note:** Version bump only for package @radixdlt/application
+
+
+
+
+
+## [1.0.13](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/application@1.0.12...@radixdlt/application@1.0.13) (2021-04-08)
+
+**Note:** Version bump only for package @radixdlt/application
+
+
+
+
+
+## [1.0.12](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/application@1.0.11...@radixdlt/application@1.0.12) (2021-04-08)
+
+**Note:** Version bump only for package @radixdlt/application
+
+
+
+
+
 ## [1.0.11](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/application@1.0.10...@radixdlt/application@1.0.11) (2021-04-02)
 
 **Note:** Version bump only for package @radixdlt/application